Lung1/samples/lung/requirements.txt
Cython
numpy
scipy
Pillow
matplotlib
scikit-image
tensorflow-gpu==1.5.0
keras==2.1.4
opencv-python
h5py
imgaug
IPython[all]
